6625-db-index-improvement

+++ db/warehouse/migrate/20241010005805_prune_warehouse_indexes_phase_1.rb
@@ -32,16 +32,16 @@ class PruneWarehouseIndexesPhase1 < ActiveRecord::Migration[7.0]
 
     raise "expected index #{index_name} to be on #{table_name} but was #{result['relname']}" unless result && result['relname'] == table_name
 
-    return unless result['scan_count'].to_i.positive?
+    return true if result['scan_count'].to_i.zero?
 
-    warn "Index #{index_name} on #{table_name} is in use (scan count > 0)."
+    Sentry.capture_message("Index #{index_name} on #{table_name} is in use (scan count > 0).")
+    false
   end
 
   # it seems that the index names aren't consistent between prod and dev. Attempt to remove the index by name only
   def remove_index_safely(table:, index_name:, columns:)
     if index_exists?(table, columns, name: index_name)
-      check_unused_index_table(index_name, table) # additional safety check
-      remove_index(table, name: index_name)
+      remove_index(table, name: index_name) if check_unused_index_table(index_name, table) # additional safety check
     # we could try and remove equivalent index but there's less certainty that is is unused if the name doesn't match
     # elsif Rails.env.development? && index_exists?(table, columns)
     #   remove_index(table, column: columns)
